Added blogger table
