+++ frontend/markbyte/package.json
@@ -13,6 +13,7 @@
     "@emotion/styled": "^11.14.0",
     "@headlessui/react": "^2.2.0",
     "@mui/material": "^6.4.3",
+    "@tanstack/react-table": "^8.21.2",
     "axios": "^1.7.9",
     "bootstrap": "^5.3.3",
     "react": "^18.3.1",
