+++ frontend/markbyte/src/data/blogdata1.json
@@ -1,0 +1,30 @@
+[
+    {
+      "title": "My First Blog Post",
+      "date": "2025-02-01T10:00:00Z",
+      "link": "#",
+      "latestVersion": "2.0",
+      "version": ["1.0", "1.1", "1.2", "2.0"],
+      "publishStatus": true,
+      "edit": "Edit"
+    },
+    {
+      "title": "Exploring TypeScript",
+      "date": "2025-02-05T14:00:00Z",
+      "link": "#",
+      "latestVersion": "2.2",
+      "version": ["1.0", "1.1", "1.2", "2.0", "2.1", "2.2"],
+      "publishStatus": false,
+      "edit": "Edit"
+    },
+    {
+      "title": "React vs Vue",
+      "date": "2025-02-10T08:00:00Z",
+      "link": "#",
+      "latestVersion": "1.0",
+      "version": ["1.0", "1.1", "1.2", "2.0"],
+      "publishStatus": true,
+      "edit": "Edit"
+    }
+  ]
+
